requirements.txt
# Content Engine API Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Image processing
Pillow>=10.0.0

# Environment variables (optional but recommended)
python-dotenv>=1.0.0

# Production server (optional)
gunicorn>=21.0.0

# Type hints support (for development)
# pydantic is included with fastapi
